Added Data Science with Python
I created a list for resources for data science with Python. It includes not only libraries, but also links to tutorials, code snippets, blog posts and talks. The existing awesome list focues on papers, my list focuses on mainly on packages and is quite different. Think it would be a good addition to this repo.